Refactor: 커뮤니티/마이페이지 네비를 공용 컴포넌트로 통합
