+++ src/app/layout.tsx
@@ -2,9 +2,10 @@
 import "./globals.css";
 import { usePathname } from "next/navigation";
 
+import CommonMainNav from "@/components/common/nav/CommonMainNav";
+import CommonSubNav from "@/components/common/nav/CommonSubNav";
 import Header from "@/components/layout/Header";
-import CommunityNav from "@/features/community/CommunityNav";
-import { MyPageMainNav, MyPagePostNav, MyPageProfileNav } from "@/features/mypage/MyPageNav";
+import MyPageProfileNav from "@/features/mypage/ProfileNav";
 import MSWProvider from "@/providers/MSWProvider";
 import QueryProvider from "@/providers/QueryProvider";
 import ReduxProvider from "@/shared/store/ReduxProvider";
@@ -23,11 +24,11 @@ export default function RootLayout({
           <QueryProvider>
             <MSWProvider>
               <Header />
-              {pathname.includes("community") && <CommunityNav />}
-              {pathname.includes("mypage") && <MyPageMainNav />}
+              {pathname.includes("community") && <CommonMainNav page="community" />}
+              {pathname.includes("mypage") && <CommonSubNav />}
               {pathname.includes("mypage/profile") && <MyPageProfileNav />}
-              {pathname.includes("mypage/posts") && <MyPagePostNav />}
-              <main className="m-auto w-full max-w-layout py-20">{children}</main>
+              {pathname.includes("mypage/posts") && <CommonMainNav page="mypage" />}
+              <main className="m-auto w-full max-w-layout">{children}</main>
             </MSWProvider>
           </QueryProvider>
         </ReduxProvider>
